[ALL CHARTS] Fixed the multiple readings of the same type flaw so that charts can be displayed for all readings types.  Also made readings into doubles instead of Integers.  Deleted old Dashboard view controller and other duplicate files.

+++ Smart Garden/Reading.swift
@@ -14,7 +14,7 @@ class Reading : PFObject, PFSubclassing {
     
     // 2
     @NSManaged var gardenID: PFObject?
-    @NSManaged var readings: Array<Array<Int>>
+    @NSManaged var readings: Array<Array<Double>>
     
     
     //MARK: PFSubclassing Protocol
